requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
polars>=0.20.0
fastexcel>=0.9.0
openpyxl>=3.1.0
xlrd>=2.0.0

# Database connectivity
# Redshift for source data
redshift-connector>=2.0.0
psycopg2-binary>=2.9.0

# MySQL for result storage
mysql-connector-python>=8.0.0

# Environment variables
python-dotenv>=1.0.0
